move chat_templates.py
