+++ tests/unit/data/hf_datasets/test_prompt_response.py
@@ -18,10 +18,8 @@ import tempfile
 import pytest
 from transformers import AutoTokenizer
 
-from nemo_rl.data.hf_datasets.chat_templates import COMMON_CHAT_TEMPLATES
-from nemo_rl.data.hf_datasets.prompt_response_dataset import (
-    PromptResponseDataset,
-)
+from nemo_rl.data.chat_templates import COMMON_CHAT_TEMPLATES
+from nemo_rl.data.hf_datasets.prompt_response_dataset import PromptResponseDataset
 
 
 @pytest.fixture
